translate: split helper fns to their own file
